Remove demo data grid
